chickenglass/requirements.txt
pyyaml~=6.0
pandoc~=2.3
flask~=2.3.3
watchdog~=3.0.0